Add files via upload
Reads an integer value from the user  then asks the user if they want to set, unset, or flip a bit inside this number. No matter which of the three options the user selects, read in another integer (in the range 0 to 31) to determine which bit to change. Finally, print out the number after the single bit change has happened.